+++ docs/changelog.rst
@@ -10,6 +10,12 @@ Changed
 - Handlers now take a single context argument (``StructurerContext`` or ``UnstructurerContext``) which contain the structurer and the target type. (PR_20_)
 
 
+Added
+^^^^^
+
+- Support for default factory in dataclasses. (PR_21_)
+
+
 Fixed
 ^^^^^
 
@@ -17,6 +23,7 @@ Fixed
 
 
 .. _PR_20: https://github.com/fjarri-eth/compages/pull/20
+.. _PR_21: https://github.com/fjarri-eth/compages/pull/21
 
 
 0.3.0 (2024-03-15)